refactor: split the workspace analytics page into sections
